[code cleaning] Mainly limit duplicated code
