+++ leshan-tl-cf-client-coap/src/main/java/org/eclipse/leshan/transport/californium/client/endpoint/coaps/CoapsClientEndpointFactory.java
@@ -24,31 +24,18 @@ import java.security.cert.X509Certificate;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.security.auth.x500.X500Principal;
-
-import org.eclipse.californium.core.coap.Message;
-import org.eclipse.californium.core.coap.Request;
 import org.eclipse.californium.core.network.CoapEndpoint;
 import org.eclipse.californium.core.network.CoapEndpoint.Builder;
-import org.eclipse.californium.core.network.Endpoint;
-import org.eclipse.californium.elements.AddressEndpointContext;
 import org.eclipse.californium.elements.Connector;
-import org.eclipse.californium.elements.DtlsEndpointContext;
-import org.eclipse.californium.elements.EndpointContext;
 import org.eclipse.californium.elements.EndpointContextMatcher;
-import org.eclipse.californium.elements.MapBasedEndpointContext;
-import org.eclipse.californium.elements.MapBasedEndpointContext.Attributes;
 import org.eclipse.californium.elements.PrincipalEndpointContextMatcher;
-import org.eclipse.californium.elements.auth.PreSharedKeyIdentity;
 import org.eclipse.californium.elements.auth.RawPublicKeyIdentity;
-import org.eclipse.californium.elements.auth.X509CertPath;
 import org.eclipse.californium.elements.config.Configuration;
 import org.eclipse.californium.elements.util.CertPathUtil;
 import org.eclipse.californium.scandium.DTLSConnector;
 import org.eclipse.californium.scandium.config.DtlsConfig;
 import org.eclipse.californium.scandium.config.DtlsConfig.DtlsRole;
 import org.eclipse.californium.scandium.config.DtlsConnectorConfig;
-import org.eclipse.californium.scandium.dtls.DtlsHandshakeTimeoutException;
 import org.eclipse.californium.scandium.dtls.cipher.CipherSuite;
 import org.eclipse.californium.scandium.dtls.pskstore.SinglePskStore;
 import org.eclipse.californium.scandium.dtls.x509.CertificateVerifier;
@@ -56,25 +43,17 @@ import org.eclipse.californium.scandium.dtls.x509.SingleCertificateProvider;
 import org.eclipse.californium.scandium.dtls.x509.StaticCertificateVerifier;
 import org.eclipse.leshan.client.endpoint.ClientEndpointToolbox;
 import org.eclipse.leshan.client.security.CertificateVerifierFactory;
-import org.eclipse.leshan.client.servers.LwM2mServer;
 import org.eclipse.leshan.client.servers.ServerInfo;
 import org.eclipse.leshan.core.SecurityMode;
 import org.eclipse.leshan.core.endpoint.EndpointUri;
 import org.eclipse.leshan.core.endpoint.Protocol;
-import org.eclipse.leshan.core.peer.IpPeer;
-import org.eclipse.leshan.core.peer.LwM2mPeer;
-import org.eclipse.leshan.core.peer.PskIdentity;
-import org.eclipse.leshan.core.peer.RpkIdentity;
-import org.eclipse.leshan.core.peer.X509Identity;
-import org.eclipse.leshan.core.request.exception.TimeoutException;
-import org.eclipse.leshan.core.request.exception.TimeoutException.Type;
-import org.eclipse.leshan.core.security.certificate.util.X509CertUtil;
 import org.eclipse.leshan.core.security.certificate.verifier.X509CertificateVerifier;
-import org.eclipse.leshan.transport.californium.DefaultExceptionTranslator;
+import org.eclipse.leshan.transport.californium.DefaultCoapsExceptionTranslator;
 import org.eclipse.leshan.transport.californium.ExceptionTranslator;
 import org.eclipse.leshan.transport.californium.Lwm2mEndpointContextMatcher;
 import org.eclipse.leshan.transport.californium.client.CaliforniumConnectionController;
 import org.eclipse.leshan.transport.californium.client.endpoint.coap.CoapClientEndpointFactory;
+import org.eclipse.leshan.transport.californium.identity.DefaultCoapsIdentityHandler;
 import org.eclipse.leshan.transport.californium.identity.IdentityHandler;
 import org.eclipse.leshan.transport.californium.security.LwM2mCertificateVerifier;
 import org.slf4j.Logger;
@@ -84,7 +63,6 @@ public class CoapsClientEndpointFactory extends CoapClientEndpointFactory {
 
     private static final Logger LOG = LoggerFactory.getLogger(CoapsClientEndpointFactory.class);
 
-    protected final String loggingTagPrefix;
     protected final CertificateVerifierFactory certificateVerifierFactory = new CertificateVerifierFactory();
 
     public CoapsClientEndpointFactory() {
@@ -97,7 +75,7 @@ public class CoapsClientEndpointFactory extends CoapClientEndpointFactory {
     }
 
     public CoapsClientEndpointFactory(String loggingTagPrefix) {
-        this.loggingTagPrefix = loggingTagPrefix;
+        super(loggingTagPrefix);
     }
 
     @Override
@@ -138,9 +116,7 @@ public class CoapsClientEndpointFactory extends CoapClientEndpointFactory {
         }
 
         // create CoAP endpoint
-        CoapEndpoint endpoint = createEndpointBuilder(dtlsConfig, defaultConfiguration, toolbox).build();
-
-        return endpoint;
+        return createEndpointBuilder(dtlsConfig, defaultConfiguration, toolbox).build();
     }
 
     protected DtlsConnectorConfig.Builder createRootDtlsConnectorConfigBuilder(Configuration configuration) {
@@ -305,77 +281,21 @@ public class CoapsClientEndpointFactory extends CoapClientEndpointFactory {
 
     @Override
     public IdentityHandler createIdentityHandler() {
-        return new IdentityHandler() {
-
-            @Override
-            public LwM2mPeer getIdentity(Message receivedMessage) {
-                EndpointContext context = receivedMessage.getSourceContext();
-                InetSocketAddress peerAddress = context.getPeerAddress();
-                Principal senderIdentity = context.getPeerIdentity();
-                if (senderIdentity != null) {
-                    if (senderIdentity instanceof PreSharedKeyIdentity) {
-                        return new IpPeer(peerAddress,
-                                new PskIdentity(((PreSharedKeyIdentity) senderIdentity).getIdentity()));
-                    } else if (senderIdentity instanceof RawPublicKeyIdentity) {
-                        PublicKey publicKey = ((RawPublicKeyIdentity) senderIdentity).getKey();
-                        return new IpPeer(peerAddress, new RpkIdentity(publicKey));
-                    } else if (senderIdentity instanceof X500Principal || senderIdentity instanceof X509CertPath) {
-                        // Extract common name
-                        String x509CommonName = X509CertUtil.extractCN(senderIdentity.getName());
-                        return new IpPeer(peerAddress, new X509Identity(x509CommonName));
-                    }
-                    throw new IllegalStateException(
-                            String.format("Unable to extract sender identity : unexpected type of Principal %s [%s]",
-                                    senderIdentity.getClass(), senderIdentity.toString()));
-                }
-                return null;
-            }
-
-            @Override
-            public EndpointContext createEndpointContext(LwM2mPeer client, boolean allowConnectionInitiation) {
-                Principal peerIdentity = null;
-                if (client.getIdentity() instanceof PskIdentity) {
-                    peerIdentity = new PreSharedKeyIdentity(((PskIdentity) client.getIdentity()).getPskIdentity());
-                } else if (client.getIdentity() instanceof RpkIdentity) {
-                    peerIdentity = new RawPublicKeyIdentity(((RpkIdentity) client.getIdentity()).getPublicKey());
-                } else if (client.getIdentity() instanceof X509Identity) {
-                    /* simplify distinguished name to CN= part */
-                    peerIdentity = new X500Principal("CN=" + ((X509Identity) client.getIdentity()).getX509CommonName());
-                } else {
-                    throw new IllegalStateException(String.format("Unsupported Identity : %s", client.getIdentity()));
-                }
-                if (client instanceof IpPeer) {
-                    IpPeer ipClient = (IpPeer) client;
-                    if (peerIdentity != null && allowConnectionInitiation) {
-                        return new MapBasedEndpointContext(ipClient.getSocketAddress(), ipClient.getVirtualHost(),
-                                peerIdentity, new Attributes().add(DtlsEndpointContext.KEY_HANDSHAKE_MODE,
-                                        DtlsEndpointContext.HANDSHAKE_MODE_AUTO));
-                    }
-                    return new AddressEndpointContext(ipClient.getSocketAddress(), ipClient.getVirtualHost(),
-                            peerIdentity);
-
-                } else {
-                    throw new IllegalStateException(String.format("Unsupported peer : %s", client));
-                }
-            }
-        };
+        return new DefaultCoapsIdentityHandler(true);
 
     }
 
     @Override
     public CaliforniumConnectionController createConnectionController() {
-        return new CaliforniumConnectionController() {
-            @Override
-            public void forceReconnection(Endpoint endpoint, LwM2mServer server, boolean resume) {
-                Connector connector = ((CoapEndpoint) endpoint).getConnector();
-                if (connector instanceof DTLSConnector) {
-                    if (resume) {
-                        LOG.info("Clear DTLS session for resumption for server {}", server.getUri());
-                        ((DTLSConnector) connector).forceResumeAllSessions();
-                    } else {
-                        LOG.info("Clear DTLS session for server {}", server.getUri());
-                        ((DTLSConnector) connector).clearConnectionState();
-                    }
+        return (endpoint, server, resume) -> {
+            Connector connector = ((CoapEndpoint) endpoint).getConnector();
+            if (connector instanceof DTLSConnector) {
+                if (resume) {
+                    LOG.info("Clear DTLS session for resumption for server {}", server.getUri());
+                    ((DTLSConnector) connector).forceResumeAllSessions();
+                } else {
+                    LOG.info("Clear DTLS session for server {}", server.getUri());
+                    ((DTLSConnector) connector).clearConnectionState();
                 }
             }
         };
@@ -383,16 +303,6 @@ public class CoapsClientEndpointFactory extends CoapClientEndpointFactory {
 
     @Override
     public ExceptionTranslator createExceptionTranslator() {
-        return new DefaultExceptionTranslator() {
-            @Override
-            public Exception translate(Request coapRequest, Throwable error) {
-                if (error instanceof DtlsHandshakeTimeoutException) {
-                    return new TimeoutException(Type.DTLS_HANDSHAKE_TIMEOUT, error,
-                            "Request %s timeout : dtls handshake timeout", coapRequest.getURI());
-                } else {
-                    return super.translate(coapRequest, error);
-                }
-            }
-        };
+        return new DefaultCoapsExceptionTranslator();
     }
 }
